refactor(timeline): extract useScrollLeft hook and TimeScaleContainer
