+++ src/pages/EditionView/tabs/ScheduleTab/horizontal/TimelineContainer.tsx
@@ -1,5 +1,5 @@
-import { useLayoutEffect, useRef, useState } from "react";
-import { TimeScale } from "./TimeScale";
+import { useRef, useState } from "react";
+import { TimeScaleContainer } from "./TimeScaleContainer";
 import { StageRow } from "./StageRow";
 import { StageLabels } from "./StageLabels";
 import { TimelineToolbar } from "./TimelineToolbar";
@@ -15,6 +15,7 @@ import type { ScheduleDay } from "@/hooks/useScheduleData";
 import { useTimelineScrollSync } from "@/hooks/useTimelineScrollSync";
 import { jumpToTimelineMoment } from "@/lib/timelineDayJump";
 import { useActiveTimelineDay } from "@/hooks/useActiveTimelineDay";
+import { useScrollLeft } from "./useScrollLeft";
 import { HEADER_STRIP_TOP_PX } from "@/lib/layout-constants";
 import { useIsMobile } from "@/hooks/use-mobile";
 
@@ -37,7 +38,7 @@ export function TimelineContainer({
 }: TimelineContainerProps) {
   const scrollContainerRef = useRef<HTMLDivElement>(null);
   const [isOverviewExpanded, setIsOverviewExpanded] = useState(false);
-  const [scrollLeft, setScrollLeft] = useState(0);
+  const scrollLeft = useScrollLeft(scrollContainerRef);
   const isMobile = useIsMobile();
   const headerStripTop = isMobile
     ? HEADER_STRIP_TOP_PX.mobile
@@ -51,22 +52,6 @@ export function TimelineContainer({
     now,
   });
 
-  useLayoutEffect(() => {
-    const container = scrollContainerRef.current;
-    if (!container) return;
-
-    function handleScroll() {
-      setScrollLeft(container!.scrollLeft);
-    }
-
-    // Read the initial position after useTimelineScrollSync's own mount
-    // scroll (also a layout effect, run before this one) so the header
-    // strip doesn't flash at translateX(0) before the first scroll event.
-    setScrollLeft(container.scrollLeft);
-
-    container.addEventListener("scroll", handleScroll, { passive: true });
-    return () => container.removeEventListener("scroll", handleScroll);
-  }, []);
   const activeDay = useActiveTimelineDay({
     scrollContainerRef,
     days: scheduleDays,
@@ -112,24 +97,12 @@ export function TimelineContainer({
           onJump={(moment) => jumpTo(moment, "center")}
         />
       )}
-      <div
-        className="sticky z-30 overflow-hidden bg-gray-900/95 backdrop-blur-md"
-        style={{ top: headerStripTop }}
-      >
-        <div
-          style={{
-            transform: `translateX(-${scrollLeft}px)`,
-            width: timelineData.totalWidth,
-          }}
-        >
-          <TimeScale
-            timeSlots={timelineData.timeSlots}
-            totalWidth={timelineData.totalWidth}
-            timezone={timezone}
-            scrollLeft={scrollLeft}
-          />
-        </div>
-      </div>
+      <TimeScaleContainer
+        timelineData={timelineData}
+        timezone={timezone}
+        scrollLeft={scrollLeft}
+        top={headerStripTop}
+      />
 
       <div className="relative">
         <StageLabels stages={timelineData.stages} />
